:sparkles: Bug fix slashing event filtering
Kujira has a bug where the slashing event is split in 2.
One of the 2 events contains 3 attributes: "address", "power" and "reason",
but not the "jailed" address.
The other one contains only 1 attribute: the "jailed" address.
Here we are filtering to keep only the former one that contains more info.
This was the error:
```
src/indexer.ts:148
  assert.ok(address && rawPower && reason);
         ^
AssertionError [ERR_ASSERTION]: undefined == true
    at decodeSlashEvent (src/indexer.ts:148:10)
    at src/indexer.ts:157:35
    at Array.map (<anonymous>)
    at decodeSlashEvents (src/indexer.ts:157:15)
    at src/indexer.ts:163:32
    at Array.forEach (<anonymous>)
    at logDecodeSlashEvents (src/indexer.ts:161:16)
    at src/indexer.ts:205:3
    at step (src/indexer.ts:67:23)
    at Object.next (src/indexer.ts:48:53) {
  generatedMessage: true,
  code: 'ERR_ASSERTION',
  actual: undefined,
  expected: true,
  operator: '=='
}
```